moved nba_reference to nba_data and trimmed down imports
